backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.2
drf-spectacular==0.27.2
dj-database-url==2.1.0
psycopg2-binary==2.9.9
redis==5.0.4
django-redis==5.4.0
celery==5.3.6
razorpay==1.4.1
twilio==8.13.0
sendgrid==6.11.0
Pillow==10.3.0
whitenoise==6.6.0
gunicorn==21.2.0
python-decouple==3.8
cloudinary==1.39.1
django-cloudinary-storage==0.3.0
